+++ examples/foundational/07x-interruptible-gnani.py
@@ -1,33 +1,23 @@
 #
-# Copyright (c) 2024–2025, Daily
+# Copyright (c) 2024-2026, Daily
 #
 # SPDX-License-Identifier: BSD 2-Clause License
 #
 
-"""
-Example demonstrating Gnani STT service integration with Pipecat.
-
-This example shows how to use Gnani's multilingual speech-to-text service
-in a conversational AI pipeline with interruption support.
-"""
 
 import os
 
 from dotenv import load_dotenv
 from loguru import logger
-from pipecat.turns.turn_strategies import (
-    TurnAnalyzerBotTurnStartStrategy,
-    TurnStartStrategies,
-)
-from pipecat.turns.turn_utils import LocalSmartTurnAnalyzerV3
-from pipecat.universal_context.llm_context import LLMContext
 
+from pipecat.audio.turn.smart_turn.local_smart_turn_v3 import LocalSmartTurnAnalyzerV3
 from pipecat.audio.vad.silero import SileroVADAnalyzer
 from pipecat.audio.vad.vad_analyzer import VADParams
 from pipecat.pipeline.pipeline import Pipeline
 from pipecat.pipeline.runner import PipelineRunner
-from pipecat.pipeline.task import PipelineTask
-from pipecat.processors.aggregators.openai_llm_context import (
+from pipecat.pipeline.task import PipelineParams, PipelineTask
+from pipecat.processors.aggregators.llm_context import LLMContext
+from pipecat.processors.aggregators.llm_response_universal import (
     LLMContextAggregatorPair,
     LLMUserAggregatorParams,
 )
@@ -36,9 +26,12 @@ from pipecat.runner.utils import create_transport
 from pipecat.services.cartesia.tts import CartesiaTTSService
 from pipecat.services.gnani.stt import GnaniSTTService
 from pipecat.services.openai.llm import OpenAILLMService
+from pipecat.transcriptions.language import Language
 from pipecat.transports.base_transport import BaseTransport, TransportParams
 from pipecat.transports.daily.transport import DailyParams
 from pipecat.transports.websocket.fastapi import FastAPIWebsocketParams
+from pipecat.turns.user_stop import TurnAnalyzerUserTurnStopStrategy
+from pipecat.turns.user_turn_strategies import UserTurnStrategies
 
 load_dotenv(override=True)
 
@@ -68,36 +61,28 @@ transport_params = {
 async def run_bot(transport: BaseTransport, runner_args: RunnerArguments):
     logger.info("Starting bot with Gnani STT")
 
-    # Initialize Gnani STT service with API credentials
-    # You can set the language to any supported Indian language
-    from pipecat.transcriptions.language import Language
-
     stt = GnaniSTTService(
         api_key=os.getenv("GNANI_API_KEY"),
         organization_id=os.getenv("GNANI_ORGANIZATION_ID"),
         params=GnaniSTTService.InputParams(
-            language=Language.HI_IN,  # Hindi by default, change as needed
+            language=Language.EN_IN,
             api_user_id=os.getenv("GNANI_USER_ID", "pipecat-user"),
         ),
     )
 
-    # TTS service
     tts = CartesiaTTSService(
         api_key=os.getenv("CARTESIA_API_KEY"),
-        voice_id="71a7ad14-091c-4e8e-a314-022ece01c121",  # British Reading Lady
+        voice_id="71a7ad14-091c-4e8e-a314-022ece01c121",
     )
 
-    # LLM service
     llm = OpenAILLMService(api_key=os.getenv("OPENAI_API_KEY"))
 
-    # System message for the LLM
     messages = [
         {
             "role": "system",
-            "content": "You are a helpful AI assistant that can understand and respond in multiple Indian languages. "
-            "Your goal is to demonstrate your capabilities in a succinct way. "
-            "Your output will be spoken aloud, so avoid special characters that can't easily be spoken. "
-            "Respond to what the user says in a creative and helpful way.",
+            "content": "You are a friendly AI assistant. Respond naturally and keep "
+            "your answers conversational. Your output will be spoken aloud, "
+            "so avoid special characters that can't easily be spoken.",
         },
     ]
 
@@ -105,35 +90,41 @@ async def run_bot(transport: BaseTransport, runner_args: RunnerArguments):
     context_aggregator = LLMContextAggregatorPair(
         context,
         user_params=LLMUserAggregatorParams(
-            turn_start_strategies=TurnStartStrategies(
-                bot=[TurnAnalyzerBotTurnStartStrategy(turn_analyzer=LocalSmartTurnAnalyzerV3())]
+            user_turn_strategies=UserTurnStrategies(
+                stop=[TurnAnalyzerUserTurnStopStrategy(turn_analyzer=LocalSmartTurnAnalyzerV3())],
             ),
         ),
     )
 
     pipeline = Pipeline(
         [
-            transport.input(),  # Transport user input
-            stt,  # Gnani STT
-            context_aggregator.user(),  # User responses
-            llm,  # LLM
-            tts,  # TTS
-            transport.output(),  # Transport bot output
-            context_aggregator.assistant(),  # Assistant spoken responses
+            transport.input(),
+            stt,
+            context_aggregator.user(),
+            llm,
+            tts,
+            transport.output(),
+            context_aggregator.assistant(),
         ]
     )
 
     task = PipelineTask(
         pipeline,
+        params=PipelineParams(
+            enable_metrics=True,
+            enable_usage_metrics=True,
+        ),
         idle_timeout_secs=runner_args.pipeline_idle_timeout_secs,
     )
 
-    # Register an event handler for when a client connects
     @transport.event_handler("on_client_connected")
     async def on_client_connected(transport, client):
         logger.info("Client connected")
-        # You can greet the user when they connect
-        # await task.queue_frames([TTSSpeakFrame("Hello! I'm using Gnani STT service. How can I help you?")])
+
+    @transport.event_handler("on_client_disconnected")
+    async def on_client_disconnected(transport, client):
+        logger.info("Client disconnected")
+        await task.cancel()
 
     runner = PipelineRunner(handle_sigint=runner_args.handle_sigint)
 
@@ -147,8 +138,6 @@ async def bot(runner_args: RunnerArguments):
 
 
 if __name__ == "__main__":
-    import asyncio
+    from pipecat.runner.run import main
 
-    from pipecat.runner import main as run
-
-    asyncio.run(run(bot))
+    main()
